requirements.txt
### dependency
tqdm
pandas
numpy
PyYAML
tensorflow-gpu==2.4.0
### local packages -
-e . # << uncomment this to install from setup.py
